Visual Display of Diff Order Polynomial Regression
The nature of distance changes in high dimensions, as evidenced in the derivation of the Maxwell distribution in statistical physics where the fact that all the volume of a d-dimensional sphere of radius r is contained in a small spherical shell around r is exploited. Almost all critical points of a function (i.e., the points where all derivatives vanish) are saddles rather than maxima or minima (an observation first made in physics in the context of the p-spin spherical spin glass). For all these reasons, it turns out that for complicated models studied in ML, predicting and fitting are very different things

Mehta, P., Bukov, M., Wang, C., Day, A., Richardson, C., Fisher, C. and Schwab, D., 2019. A high-bias, low-variance introduction to Machine Learning for physicists. Physics Reports, 810, pp.1-124. (Acessed 18 July 2020)